Fix review notes
