+++ ydb/core/formats/arrow/ut/ya.make
@@ -34,7 +34,6 @@ SRCS(
     ut_column_filter.cpp
     ut_hash.cpp
     ut_reader.cpp
-    ut_reverse.cpp
 )
 
 END()
