+++ ydb/tests/olap/delete/test_delete_all_after_inserts.py
@@ -51,40 +51,54 @@ class TestDeleteAllAfterInserts(DeleteTestBase):
 
         assert self._get_row_count(table_path) == 0
 
-        # passes
         assert len(self.ydb_client.query(
             f"""
-            SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000 ORDER by id1, id2 LIMIT 100;
+            SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000 ORDER by id1 ASC, id2 ASC LIMIT 100;
             """
         )[0].rows) == 0
 
-        # passes
+        assert len(self.ydb_client.query(
+            f"""
+            SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000 ORDER by id1 ASC, id2 ASC;
+            """
+        )[0].rows) == 0
+
         assert len(self.ydb_client.query(
             f"""
             SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000 ORDER by id1 DESC, id2 DESC;
             """
         )[0].rows) == 0
 
-        # passes
         assert len(self.ydb_client.query(
             f"""
             SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000 LIMIT 100;
             """
         )[0].rows) == 0
 
-        # passes
         assert len(self.ydb_client.query(
             f"""
             SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000;
             """
         )[0].rows) == 0
 
-        # fails
         assert len(self.ydb_client.query(
             f"""
             SELECT id1 as id1, id2 as id2 FROM `{table_path}` WHERE id1 != 10000000 ORDER by id1 DESC, id2 DESC LIMIT 100;
             """
         )[0].rows) == 0
+
+    def _test_select_query(self, query, reversed=False):
+        res = self.ydb_client.query(query)[0].rows
+        prev_row = None
+        for i in range(len(res)):
+            row = res[i]
+            assert row["value"] == 'NEW', row["id1"]
+            if prev_row is not None:
+                if reversed:
+                    assert row["id1"] < prev_row["id1"]
+                else:
+                    assert row["id1"] < prev_row["id1"]
+            prev_row = row
 
     def _test_update_all(self, rows_to_insert, insert_iterations, all_rows_ids=None):
         table_path = f"{self._get_test_dir()}/testTableAllUpdated"
@@ -120,29 +134,33 @@ class TestDeleteAllAfterInserts(DeleteTestBase):
 
         assert self._get_row_count(table_path) == rows_in_table
 
-        # passes
-        res = self.ydb_client.query(
+        self._test_select_query(
             f"""
             SELECT id1 as id1, value as value FROM `{table_path}` WHERE id1 != 10000001 ORDER by id1 DESC;
-            """
-        )[0].rows
+            """,
+            reversed=True
+        )
 
-        for row in res:
-            assert row["value"] == 'NEW'
+        self._test_select_query(
+            f"""
+            SELECT id1 as id1, value as value FROM `{table_path}` WHERE id1 != 10000001 ORDER by id1 ASC;
+            """,
+            reversed=False
+        )
 
-        res = self.ydb_client.query(
+        self._test_select_query(
+            f"""
+            SELECT id1 as id1, value as value FROM `{table_path}` WHERE id1 != 10000001 ORDER by id1 ASC LIMIT 100;
+            """,
+            reversed=False
+        )
+
+        self._test_select_query(
             f"""
             SELECT id1 as id1, value as value FROM `{table_path}` WHERE id1 != 10000001 ORDER by id1 DESC LIMIT 100;
-            """
-        )[0].rows
-
-        prev_row = None
-        for i in range(len(res)):
-            row = res[i]
-            assert row["value"] == 'NEW', row["id1"]
-            if prev_row is not None:
-                assert row["id1"] < prev_row["id1"]
-            prev_row = row
+            """,
+            reversed=True
+        )
 
     def test_delete_all_rows_after_several_inserts(self):
         # IMPORTANT note: tests passes with 1 insert_iterations
